move IElectronBridge from electron to web
