+++ electron/tsconfig.json
@@ -14,6 +14,7 @@
     }
   },
   "include": [
-    "src/**/*"
-  ]
+    "src/**/*",
+    "../web/src/IElectronBridge.ts"
+]
 }
